Moved classes into appropriate packages, deleted redundant DeleteController
