+++ src/main/java/com/example/CustomerRepository.java
@@ -1,5 +1,6 @@
 package com.example;
 
+import com.example.Models.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository extends JpaRepository<Customer, Integer> {
