+++ src/main/java/com/example/Interfaces/IDeleteService.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.Interfaces;
 
 public interface IDeleteService {
