+++ src/main/java/com/example/Interfaces/IGetService.java
@@ -1,4 +1,6 @@
-package com.example;
+package com.example.Interfaces;
+
+import com.example.Models.Customer;
 
 import java.util.List;
