+++ src/main/java/com/example/Services/DeleteService.java
@@ -1,9 +1,11 @@
-package com.example;
+package com.example.Services;
 
+import com.example.CustomerRepository;
+import com.example.Interfaces.IDeleteService;
 import org.springframework.stereotype.Service;
 
 @Service
-public class DeleteService implements IDeleteService{
+public class DeleteService implements IDeleteService {
 
     private final CustomerRepository customerRepository;
